requirements.txt
Django==1.9.5
django-grappelli==2.8.1
django-markdown-deux==1.0.5
django-pagedown==0.1.1
markdown2==2.3.1
Pillow==3.2.0
requests==2.9.1
wheel==0.24.0
